Fixed Timestamp() to work without side effect.
The problem can be illustrated in this example:

	var observable = Observable.Generate (
			1, x => x < 6, x => x + 1, x => x,
			x=>TimeSpan.FromSeconds(1), Scheduler.ThreadPool)
		.Timestamp();
	observable.Subscribe(Console.WriteLine);
	Thread.Sleep (3000);
	observable.Subscribe(Console.WriteLine);

The second subscription did not start with 1 (3 or 4, up to run-time). It was
because Timestamp() used to use Subject<T> and the source input to that method
only gives values per second (i.e. it is a hot observable).

With this change it gives values from 1 for the second subscription as
expected.